greatly simplify code
