+++ src/conn/lnd/mod.rs
@@ -1,3 +1,4 @@
-pub mod unlocker;
 pub mod lndrpc;
+pub mod setup;
+pub mod unlocker;
 pub mod utils;
