+++ src/images/relay.rs
@@ -1,6 +1,7 @@
 use super::traefik::traefik_labels;
 use super::*;
-use crate::config::Node;
+use crate::config::{Clients, Node};
+use crate::conn::relay::setup::relay_client;
 use crate::images::lnd::to_lnd_network;
 use crate::utils::{domain, exposed_ports, host_config, volume_string};
 use anyhow::{Context, Result};
@@ -38,6 +39,15 @@ impl RelayImage {
         if let Some(h) = eh {
             self.host = Some(format!("{}.{}", self.name, h));
         }
+    }
+    pub async fn connect_client(&self, proj: &str, clients: &mut Clients) -> Result<()> {
+        match relay_client(proj, self).await {
+            Ok(client) => {
+                clients.relay.insert(self.name.clone(), client);
+            }
+            Err(e) => log::warn!("relay_client error: {:?}", e),
+        };
+        Ok(())
     }
 }
